Refactor DetailsTile
Rename to DetailsListTile.
Remove redundant InkWell.
Use named instead of positional arguments.
